Deposit and Withdrawal and modification according to rtm and api sheet

+++ src/main/java/com/training/demo/repository/CoinManagementRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.training.demo.model.CoinManagement;
 import com.training.demo.model.User;
 
-public interface CoinManagementRepository extends JpaRepository<CoinManagement,Long> {
+public interface CoinManagementRepository extends JpaRepository<CoinManagement,Integer> {
 	
-	public CoinManagement findOneByCoinId(Long coinId);
+	public CoinManagement findOneByCoinId(Integer coinId);
 	
 	public CoinManagement findOneByCoinName(String coinName);
